sweet alert

+++ src/Community/Posts/Delete.jsx
@@ -2,21 +2,40 @@ import React from 'react';
 import DeleteIcon from '@mui/icons-material/Delete';
 import { collection, getDocs, doc, deleteDoc } from 'firebase/firestore';
 import { db } from '../../firebase';
+import swal from 'sweetalert';
 
-export const Delete = ({ id, avatar, setPosts }) => {
+
+export const Delete = ({ id, setPosts }) => {
+
   const postsCollectionRef = collection(db, 'Posts');
   //Eliminar Post
   const deletePost = async (id) => {
     //console.log(id);
-    try {
-      //Eliminar
-      await deleteDoc(doc(db, 'Posts', id));
-      //console.log('Document deleted with ID: ', id);
-      //Actualizar estado
-      getAllData();
-    } catch (error) {
-      //console.error('Error adding document: ', error);
-    }
+    await swal({
+      title: "¿Estás seguro?",
+      text: "Una vez eliminado, no podrás recuperar este post.",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        try {
+          //Eliminar
+          deleteDoc(doc(db, 'Posts', id));
+          //console.log('Document deleted with ID: ', id);
+          //Actualizar estado
+          getAllData();
+        } catch (error) {
+          //console.error('Error adding document: ', error);
+        }
+        swal("¡Este post ha sido eliminado!", {
+          icon: "success",
+        });
+      } else {
+        swal("¡Uuf! ¡Salvaste este post!");
+      }
+    });
   };
 
   //Traer nueva data actualizada
